+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ts-linq/orm
 
+## 6.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/core@3.4.8
+  - @ts-linq/query@4.2.2
+  - @ts-linq/migrations@2.8.6
+
+## 6.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.9.0
+  - @ts-linq/concurrency@3.0.9
+  - @ts-linq/core@3.4.7
+  - @ts-linq/metadata@4.1.8
+  - @ts-linq/metrics-safe@1.2.11
+  - @ts-linq/migrations@2.8.5
+  - @ts-linq/query@4.2.1
+  - @ts-linq/sql-visitor@4.3.5
+  - @ts-linq/telemetry@2.3.2
+
 ## 6.0.3
 
 ### Patch Changes