+++ packages/sql-visitor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-linq/sql-visitor
 
+## 4.3.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.9.0
+  - @ts-linq/ast@3.2.9
+
 ## 4.3.4
 
 ### Patch Changes